[TASK] Bugfix company persistence
Change pushed in order to have a decent environment for the demo.
Had to cut some corners here and there, so we may want to revert some
of these changes later. An example: we should probably not remove the
notEmpty validation on addresscode, but somehow the annotation
validationGroup (on controller) does not work when editing a related entity

Resolves: #591

Change-Id: I8b21a5e588154770e6b4b38eac99579ea51c800e