[TASK] Use Font Awesome instead of sprites in Twitter Bootstrap
Removes the import of sprites into the Groeiservice Bootstrap.
Adds Fontawesome files and includes these instead of the sprites.
Contains scss sourcefiles and css files ready to use.

The sourcefiles are fixed and good to  use, but if the files are compiled
again from source, the fontawesome resourcepaths will be broken.

Change-Id: I642f5b32f205a1fbe4d812e6f39b61936b4322c0
Related: #733
Related: #844