[TASK] Restructure Controllers and Router in Emberjs
This change is required as we ended up with some loose
ends in JavaScript after the Emberjs implementation.
Now controllers can be registered in different packages and
in most cases don't even need any JavaScript code.

Todo: make sure JavaScript files are registered in the
package they belong to.

Change-Id: Idf60b7064dfd8bfc2391b430e68f03fcf188da8a
Related: #923