[TASK] Add websocket support
With this change the application will initialize a websocket connection on
page load. When the connection is lost a warning is show and it's
attempted to reinitialize the connection. If this failed another warning
is shown, and a re-attempt will follow.

It's possible to register methods to call when a message is received.

Change-Id: If1fa28097a2f8ef45831f836fd59f633759ca8e1
Resolves: #950