[TASK] Make the widget dynamic.
This change adds a controller to be used for fetching tasks
with AJAX calls. This is just a temporary solution till the sockets
can be used. Besides this the change cleans up the notification
controller which is replaced by sockets already, and changes
the webRedirect configuration along the way.

MM-11 MM-15 #resolve

Change-Id: I0dae23ea343740664779f8d98cc5d3e006e5ae57