[!!!][TASK] Replace Famelo.Gerrit by ./gerrit_update.php
This change removes the ./flow gerrit:update command and replaces
it by ./gerrit_update.php.
The ./flow gerrit:update command did not work anymore as it was
run from within Flow which means that the reflection stumbled over
missing classes and interfaces which should be pulled in by the
command.

Change-Id: I503fa3e0a674932b417deefdbb5c598a2b48c203