[BUGFIX] The chosen elements do not use the correct font
Bootstrap variables overrule the theme variable and the
chosen css does not respect the bootstrap styles.

This patch moves the chosen css to the themes and merges the
bootstrap variable with theme variable.
Now bootstrap and the theme use the same sass variables.

Choses styles are cleaned-up and better respect bootstrap styles.

Futher some cleanup of not used files.
re-test
Resolves: MM-340

Change-Id: I66a6a765499e4bf6280140455c76ac8406431565